A "short" in the code to place workers for a city needed to be replaced with an int. (missed in the patch to increase the number of players from 14 to 30). I checked the rest of the code once again for missed shorts.
[[originally from CVS; cvs2svn created svn r2279]]

+++ server/cityturn.c
@@ -48,7 +48,7 @@
 
 #include "cityturn.h"
 
-extern signed short int minimap[MAP_MAX_WIDTH][MAP_MAX_HEIGHT];
+extern signed int minimap[MAP_MAX_WIDTH][MAP_MAX_HEIGHT];
 
 static void set_tax_income(struct city *pcity);
 static void set_food_trade_shields(struct city *pcity);
